# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Fidelito API
  version: 2.0.0
  description: |
    Digital loyalty card platform for Tunisian businesses.
    Clients display dynamic TOTP QR codes; cashiers scan to stamp loyalty cards.
    Businesses manage staff, rewards, and analytics. Admins manage merchants.
servers:
  - url: http://localhost:8787
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_token
  schemas:
    Profile:
      type: object
      properties:
        id: { type: string }
        role: { type: string, enum: [client, cashier, business, admin] }
        firstName: { type: string, nullable: true }
        lastName: { type: string, nullable: true }
        email: { type: string }
        address: { type: string, nullable: true }
        locale: { type: string, enum: [en, fr, ar] }
        createdAt: { type: string }
        lastUpdated: { type: string }
...
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /api/auth/otp/send-verification:
    post:
      summary: Send email verification code
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string }
                locale: { type: string, enum: [en, fr, ar] }
      responses:
        '200': { description: Sent }
  /api/auth/otp/verify-email:
    post:
      summary: Verify email with code
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [email, code]
              properties:
                email: { type: string }
                code: { type: string }
      responses:
        '200': { description: Verified }

  # --- Client ---
...
  /api/me:
    get:
      summary: Get own profile
      responses:
        '200':
          description: Profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  profile: { $ref: '#/components/schemas/Profile' }
    patch:
      summary: Update own profile
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName: { type: string }
                lastName: { type: string }
                address: { type: string, nullable: true }
                locale: { type: string, enum: [en, fr, ar] }
      responses:
        '200': { description: Updated profile }

  # --- Admin ---
...
  /api/admin/users/{id}:
    get:
      summary: Get user by ID
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: User }
    patch:
      summary: Update user
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                role: { type: string, enum: [client, cashier, business, admin] }
                firstName: { type: string }
                lastName: { type: string }
                address: { type: string, nullable: true }
                locale: { type: string, enum: [en, fr, ar] }
      responses:
        '200': { description: Updated }
    delete:
      summary: Delete user
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Deleted }
...
  /api/admin/merchants:
    get:
      summary: List all merchants
      responses:
        '200':
          description: Merchants
    post:
      summary: Create merchant + assign owner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, slug, ownerEmail]
              properties:
                name: { type: string }
                slug: { type: string }
                ownerEmail: { type: string }
                stampsPerReward: { type: integer, default: 10 }
                planTier: { type: string, enum: [starter, growth, pro], default: starter }
                monthlyPointCap: { type: integer, default: 300 }
      responses:
        '201': { description: Created }
  /api/admin/merchants/{id}:
    get:
      summary: Get merchant details
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Merchant }
    patch:
      summary: Update merchant
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                stampsPerReward: { type: integer }
                planTier: { type: string, enum: [starter, growth, pro] }
                monthlyPointCap: { type: integer }
                isActive: { type: boolean }
      responses:
        '200': { description: Updated }
    delete:
      summary: Soft-delete merchant (set inactive)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Deleted }
  /api/admin/merchants/{id}/staff:
    get:
      summary: List merchant staff
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Staff }
  /api/admin/merchants/{id}/stats:
    get:
      summary: Merchant analytics
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Stats }